pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repixelizer"
version = "0.1.0"
description = "Optimization-based repixelizer for converting fake pixel art into true grid-aligned sprite and icon art."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
  { name = "OpenAI Codex" }
]
dependencies = [
  "numpy>=2.1.0",
  "Pillow>=10.4.0",
  "torch>=2.6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0"
]

[project.scripts]
repixelize = "repixelizer.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
